feat: make decorateColumn and getBy public
